Pipeline-code/requirements.txt
python-dotenv
requests
sqlalchemy
pymysql
pandas
python-dateutil
loguru